security/pfSense-pkg-softether5: OpenSSL compatibility and persistent data config

+++ security/pfSense-pkg-softether5/files/usr/local/etc/rc.d/vpnbridge.sh
@@ -6,7 +6,7 @@ name="vpnbridge"
 command="/usr/local/bin/vpnbridge"
 
 piddir="/var/run/softether"
-datadir="/cf/softether"
+datadir="/cf/conf/softether"
 link_datadir="/var/db/softether"
 
 precmd() {
@@ -17,7 +17,7 @@ precmd() {
         mkdir -p ${datadir}
     fi
 
-    if [ ! -L "${link_datadir}" ]; then
+    if [ ! -L "${link_datadir}" ] && [ ! -e "${link_datadir}" ]; then
         ln -s ${datadir} ${link_datadir}
     fi
 }
